[update] refactor Landing
