+++ client/src/components/Landing/Landing.css
@@ -20,28 +20,33 @@
   border: none;
   border-radius: 10px;
 }
-.landing .blurbs, .landing .cards {
+.landing .blurbs,
+.landing .cards {
   display: flex;
   flex-direction: row;
   justify-content: center;
   flex-wrap: wrap;
 }
-.landing .blurbs > *, .landing .cards > * {
+.landing .blurbs > *,
+.landing .cards > * {
   display: flex;
   flex-direction: column;
   align-items: center;
   margin: 2rem;
   padding: 3rem;
 }
-.landing .blurbs > * > *, .landing .cards > * > * {
+.landing .blurbs > * > *,
+.landing .cards > * > * {
   margin: 0.5rem;
 }
-.landing .blurbs > * img, .landing .cards > * img {
+.landing .blurbs > * img,
+.landing .cards > * img {
   height: auto;
   width: 100px;
   border-radius: 50%;
 }
-.landing .blurbs > .card, .landing .cards > .card {
+.landing .blurbs > .card,
+.landing .cards > .card {
   border: 2px solid lightgray;
   border-radius: 10px;
 }
